requirements.txt
streamlit
openai
sentence-transformers
faiss-cpu
python-pptx
python-docx
pymupdf
pandas